requirements.txt
Flask==2.3.2
matplotlib==3.7.1
numpy==1.23.5
opencv_contrib_python==4.8.1.78
opencv_python==4.7.0.72
opencv_python_headless==4.8.0.74
Pillow==9.4.0
Pillow==10.2.0
pytesseract==0.3.10
Requests==2.31.0
Shapely==2.0.3
ultralytics==8.0.180
python-dotenv==1.0.0